shoppingApp/requirements.txt

asgiref~=3.8
Django~=5.2
sqlparse==0.5.0
Pillow~=11.2
flower==2.0.1
celery==5.4.0
stripe==9.3.0
python-decouple==3.8
WeasyPrint==61.2
redis==5.0.4
django-rosetta==0.10.0
django-parler==2.3
django-localflavor==4.0
